# museum-backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Musaium API",
    "version": "1.1.0",
    "description": "Active API surface for Musaium mobile backend (health, auth, chat, daily-art, reviews, admin, museum, support)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key (msk_...)",
        "description": "B2B API key authentication. Pass the full API key as a Bearer token."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "F7 (2026-04-30) — Web admin session cookie. HttpOnly + Secure + SameSite=Strict. State-changing requests authenticated via this cookie MUST also send the X-CSRF-Token header (double-submit, value bound to the access_token via HMAC-SHA256)."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "F7 — CSRF double-submit token. Required on POST/PUT/PATCH/DELETE when the request authenticates via the access_token cookie. Value MUST equal the csrf_token cookie (which equals HMAC-SHA256(access_token, CSRF_SECRET))."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "additionalProperties": true
          }
        }
      },
...
      "AuthUser": {
        "type": "object",
        "required": ["id", "email", "role", "onboardingCompleted"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstname": {
            "type": ["string", "null"]
          },
          "lastname": {
            "type": ["string", "null"]
          },
          "role": {
            "type": "string",
            "enum": ["visitor", "moderator", "museum_manager", "admin"],
            "description": "User role for RBAC"
          },
          "onboardingCompleted": {
            "type": "boolean",
            "description": "Whether the user has completed the onboarding flow"
          },
          "ttsVoice": {
            "type": "string",
            "enum": ["alloy", "echo", "fable", "onyx", "nova", "shimmer"],
            "nullable": true,
            "description": "Spec C T2.4 — visitor's preferred TTS voice. `null` means inherit the env-level default."
          }
        }
      },
      "AuthSessionResponse": {
        "type": "object",
        "required": ["accessToken", "refreshToken", "expiresIn", "refreshExpiresIn", "user"],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer"
          },
          "refreshExpiresIn": {
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "mfaWarningDaysRemaining": {
            "type": "integer",
            "minimum": 0,
            "description": "R16 — present when an admin is inside the MFA warning window. Drives the enroll-now banner."
          }
        }
      },
...
      "MfaEnrollResponse": {
        "type": "object",
        "required": ["otpauthUrl", "manualSecret", "recoveryCodes"],
        "properties": {
          "otpauthUrl": {
            "type": "string",
            "description": "otpauth:// URI suitable for QR rendering."
          },
          "manualSecret": {
            "type": "string",
            "description": "Base32 secret for manual entry into authenticator apps."
          },
          "recoveryCodes": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 10,
            "maxItems": 10,
            "description": "10 plain-text recovery codes — shown ONCE."
          }
        }
      },
      "MfaVerifyResponse": {
        "type": "object",
        "required": ["enrolledAt"],
        "properties": {
          "enrolledAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MfaRecoverySessionResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/AuthSessionResponse" },
          {
            "type": "object",
            "required": ["remainingRecoveryCodes"],
            "properties": {
              "remainingRecoveryCodes": {
                "type": "integer",
                "minimum": 0,
                "description": "Number of unused recovery codes left after this redemption."
              }
            }
          }
        ]
      },
      "SessionDTO": {
        "type": "object",
        "required": ["id", "museumMode", "createdAt", "updatedAt", "intent"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "locale": {
            "type": ["string", "null"]
          },
          "museumMode": {
            "type": "boolean"
          },
          "title": {
            "type": ["string", "null"]
          },
          "museumName": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "intent": {
            "type": "string",
            "enum": ["default", "walk"],
            "description": "Session-level intent — drives prompt strategy and walk-mode UX."
          }
        }
      },
...
      "ChatMessageDTO": {
        "type": "object",
        "required": ["id", "role", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": ["user", "assistant", "system"]
          },
          "text": {
            "type": ["string", "null"]
          },
          "imageRef": {
            "type": ["string", "null"]
          },
          "image": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ChatMessageImage"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": ["object", "null"],
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
      "CreateSessionRequest": {
        "type": "object",
        "description": "Optional metadata attached to a new chat session. Every field is optional; the server derives defaults from the authenticated user and request headers.",
        "properties": {
          "locale": {
            "type": "string",
            "description": "IETF BCP 47 locale tag (e.g. 'fr', 'en-US'). Falls back to Accept-Language."
          },
          "museumMode": {
            "type": "boolean"
          },
          "museumId": {
            "type": "integer",
            "minimum": 1
          },
          "museumName": {
            "type": "string"
          },
          "museumAddress": {
            "type": "string"
          },
          "coordinates": {
            "type": "object",
            "required": ["lat", "lng"],
            "properties": {
              "lat": {
                "type": "number",
                "minimum": -90,
                "maximum": 90
              },
              "lng": {
                "type": "number",
                "minimum": -180,
                "maximum": 180
              }
            }
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": ["session"],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionDTO"
          }
        }
      },
...
      "GetSessionResponse": {
        "type": "object",
        "required": ["session", "messages", "page"],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionDTO"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDTO"
            }
          },
          "page": {
            "type": "object",
            "required": ["nextCursor", "hasMore", "limit"],
            "properties": {
              "nextCursor": {
                "type": ["string", "null"]
              },
              "hasMore": {
                "type": "boolean"
              },
              "limit": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ListSessionsResponse": {
        "type": "object",
        "required": ["sessions", "page"],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "museumMode", "createdAt", "updatedAt", "messageCount", "intent"],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "locale": {
                  "type": ["string", "null"]
                },
                "museumMode": {
                  "type": "boolean"
                },
                "title": {
                  "type": ["string", "null"]
                },
                "museumName": {
                  "type": ["string", "null"]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "intent": {
                  "type": "string",
                  "enum": ["default", "walk"],
                  "description": "Session-level intent — drives prompt strategy and walk-mode UX."
                },
                "messageCount": {
                  "type": "integer"
                },
                "preview": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["user", "assistant", "system"]
                    }
                  }
                }
              }
            }
          },
          "page": {
            "type": "object",
            "required": ["nextCursor", "hasMore", "limit"],
            "properties": {
              "nextCursor": {
                "type": ["string", "null"]
              },
              "hasMore": {
                "type": "boolean"
              },
              "limit": {
                "type": "integer"
              }
            }
          }
        }
      },
      "DeleteSessionResponse": {
        "type": "object",
        "required": ["sessionId", "deleted"],
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/mfa/status": {
      "get": {
        "tags": ["Auth"],
        "summary": "Read the calling user's MFA enrollment state (F9 envelope).",
        "description": "Self-scoped read of the calling user's MFA status. Drives the Enable/Disable MFA UI without exposing the encrypted secret. RBAC: authenticated user only — req.user.id IS the user being read.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "MFA status envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["mfaEnrolled", "methods", "lastVerifiedAt"],
                  "properties": {
                    "mfaEnrolled": {
                      "type": "boolean",
                      "description": "True iff TOTP has been enrolled AND verified at least once."
                    },
                    "methods": {
                      "type": "array",
                      "items": { "type": "string", "enum": ["totp"] },
                      "description": "Active second-factor methods. Today only TOTP. Forward-compatible with WebAuthn / push."
                    },
                    "lastVerifiedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true,
                      "description": "ISO 8601 of the last successful TOTP / recovery verification. Null when MFA never verified."
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/mfa/enroll": {
      "post": {
        "tags": ["Auth"],
        "summary": "Start (or rotate) a TOTP enrollment for the authenticated user.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Fresh secret + recovery codes (one-time view).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MfaEnrollResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": {
            "description": "MFA is already enrolled — disable it first to re-enroll."
          },
          "429": {
            "description": "Too many enrollment attempts."
          }
        }
      }
    },
    "/api/auth/mfa/enroll/verify": {
      "post": {
        "tags": ["Auth"],
        "summary": "Confirm a freshly enrolled TOTP secret with a 6-digit code.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": { "type": "string", "minLength": 6, "maxLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enrollment confirmed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MfaVerifyResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "No pending MFA enrollment." },
          "429": { "description": "Too many verification attempts." }
        }
      }
    },
    "/api/auth/mfa/challenge": {
      "post": {
        "tags": ["Auth"],
        "summary": "Exchange mfaSessionToken + 6-digit TOTP code for a real JWT pair.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mfaSessionToken", "code"],
                "properties": {
                  "mfaSessionToken": { "type": "string" },
                  "code": { "type": "string", "minLength": 6, "maxLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT session.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthSessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "description": "Too many MFA challenge attempts." }
        }
      }
    },
    "/api/auth/mfa/recovery": {
      "post": {
        "tags": ["Auth"],
        "summary": "Exchange mfaSessionToken + recovery code for a JWT pair (consumes the code).",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mfaSessionToken", "recoveryCode"],
                "properties": {
                  "mfaSessionToken": { "type": "string" },
                  "recoveryCode": { "type": "string", "minLength": 6, "maxLength": 32 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT session + remaining recovery code count.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MfaRecoverySessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "description": "Too many recovery attempts." }
        }
      }
    },
    "/api/auth/mfa/disable": {
      "post": {
        "tags": ["Auth"],
        "summary": "Disable MFA after re-confirming the user's password.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentPassword"],
                "properties": {
                  "currentPassword": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["disabled"],
                  "properties": {
                    "disabled": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "description": "Too many disable attempts." }
        }
      }
    },
...
    "/api/chat/sessions": {
      "post": {
        "tags": ["Chat"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Chat"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of sessions per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/chat/sessions/{id}": {
      "get": {
        "tags": ["Chat"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Cursor for message pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Number of messages per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Session history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Chat"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete if empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}